projeto_sptrans/api_loader/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
idna==3.9
minio==7.2.8
pycparser==2.22
pycryptodome==3.20.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
